[GITEA] Limit amount of javascript errors being shown
- Currently an unlimited amount of errors could be shown to the user,
this can cause that if something goes wrong the user will be shown a big
wall of errors, even when Forgejo in most cases can still be used.
- Therefor limit to three errors being shown, as it's unlikely that
three seperate javascript errors needs to be shown, they are likely to
be related and the console still shows all of the errors.