Rework the recently pushed branches banner conditions
With this change, the "You pushed on branch xyz" banner will only be
displayed if the base repository has pull request enabled (rather than
the viewed one), because the banner's goal is to target the base repo.

If the base repository does not allow pull requests, then the "New pull
request" button would lead to a 404 anyway, and should be disabled.
Without the button, the message itself is rather pointless.

If the repository is not a fork, then, with this change, the message is
never displayed. To send pull requests against itself, the `/branches/`
page will list local branches, and allows opening pull requests from
there.

The idea here is that the banner is mostly useful in the context of a
fork. While the workflow of opening PRs against the same repo is used
often, it rarely needs a notification, and said notification in these
cases is more an annoyance than helpful.

Signed-off-by: Gergely Nagy <forgejo@gergo.csillger.hu>