Non-org users can be assignable to issues (Codeberg/Community#935).
A user without write access to the repo/read access to the PR can now show
in the assignees dropdown list (webui/issue), provided this user has
posted in some way to the issue. The user can then be assigned.

This change provides a fix for Codeberg/Community#935,

Note: Blocked users can not be assigned to the issue, nor do they show
in the assignees dropdown list.

Users not known to the repo/issue can also be assigned even though they
are not in the drop down list.

The assignees dropdown has been modified so that the input field can
accept a value NOT in the dropdown list.
On Enter/Add the user gets added to the list of assignees.
An error gets displayed in case the user cannot get assigned.

Note: The user entered in the input field must be an exact value. There
is no live search (as when selecting blocked users).

Once a new user has been added the user appears in the
dropdown and stays listed, even if deselected as assignee.