[ACTIONS] skip superflous pull request synchronized event
Skip a HookEventPullRequestSync event if it has the same CommitSHA as
an existing HookEventPullRequest event in the ActionRun table. A
HookEventPullRequestSync event must only create an ActionRun if the
CommitSHA is different from what it was when the PR was open.

This guards against a race that can happen when the following is done
in parallel:

* A commit C is pushed to a repo on branch B
* A pull request with head on branch B

it is then possible that the the pull request is created first,
successfully. The commit that was just pushed is not known yet but the
PR only references the repository and the B branch so it is fine.

A HookEventPullRequest event is sent to the notification queue but not
processed immediately.

The commit C is pushed and processed successfully. Since the PR
already exists and has a head that matches the branch, the head of the
PR is updated with the commit C and a HookEventPullRequestSync event
is sent to the notification queue.

The HookEventPullRequest event is processed and since the head of the
PR was to be commit C, an ActionRun with CommitSHA C is created.

The HookEventPullRequestSync event is then processed and also has a
CommitSHA equal to C.

Refs: https://codeberg.org/forgejo/forgejo/issues/2009