[BUG] Include LFS files in Git archive
- Currently, generating an archive from an archive with LFS files would
result in the LFS files not containing their actual content, but
instead having LFS meta-content.
- Make the git archive command LFS aware, by setting the required config
values for git-lfs to be called and work. This cannot be done globally,
as this is actually explicitely unset to prevent problems on repository
with LFS files.
- Implement two internal routes that's used for LFS that bypasses the
authentication checks, which in turn allows git archive to work on
private repositories that contain LFS files. The internal routes are
only accessible if you possess the `INTERNAL_SECRET_TOKEN`, which isn't
public knowledge.
- Adds basic integration test.
- Resolves #2540