Do not incorrectly discard PR updates when a commit is pushed
The guard against the race incorrectly assumes the sha of the commit
being pushed belongs to the base repository. It finds the highest
possible pull request ID from the head repository instead of looking
it up in the base repository.

Figuring out if a PR was created in the future based on the highest
index of the base repository would require collecting all of them
because there is no way to know in advance which repository may be
involved in the race.

* A Created field is added to the Issue table
* The Created field is set to the time (with nano resolution) on creation
* The decision to not update a pull request created after a commit was
  pushed is based on the time ( with nano resolution) the push update
  was handled

When the "created" column is NULL, no verification is done, pull
requests that were created before the column was created in the
database cannot be newer than the latest in memory commit push event.

Fixes the regression introduced in 036f1eddc53b31c28b2a7d808fada07f35b6f9b1
"avoid superfluous synchronized pull_request run when opening a PR"

Fixes: https://codeberg.org/forgejo/forgejo/issues/2817