[FEAT] Support meilisearch for code indexing
- Add support for meilisearch to index code, it's quite normal in terms
of hacks, it follows elasticsearch's indexing code approach quite a
bit (which should be unified at some point with bleve as well). It has
support for everything that the index interface requires, it supports
fuzzy and exact search. Getting the match in the content doesn't require
`<em>` hacks like elasticsearch.
- Requires a new filename indexer ID that isn't consistent with the
other indexers as meilisearch has strict limits on the allowed
characters.
- Add unit tests.
- Add CI support.
- Ref: https://codeberg.org/Codeberg-Infrastructure/code-search/issues/20

WIP:
- Battle test this against codeberg's code search project.