do not add text/plain inline attachments as attachments to the comment text/plain inline attachments are parsed into the body of the comment, so we do not add them as an attachment to avoid duplicates
The test for the signature has been changed.
Messages are now no longer cut off after "-", "--" or "---", which may
cause some signatures to be added to the comment, but it's better to
add a little more rather than cut off important content. Everything
after more than 4 "-" are cut off.
"--" is used as a separator between inline attachments
"---" is also often used by users to separate paragraphs