fix(actions): prevent deleted records' UUID from colliding with new records
This commit changes the code that deletes a runner so it updates the
UUID before deleting the record. The new UUID is set to 8 0xff bytes
followed by a little endian version of the record's numeric ID. Such
UUIDs cannot be created from tokens when registering runners, as the
first 16 bytes of the token are in the `[0-9a-f]` range. This should
prevent deleted runners from colliding with new records if the tokens
share the same first 16 characters.