restore old image paste behavior This restores the old behavior of pasting images. With this change all images are added with the markdown syntax instead of the html tag.
This removes the down scaling of high density images, but the image size
is always limited by the width of the comment box, often images are in
tables to show differences. The markdwon syntax has the benefit that an
a tag is added around the image to open it in a new tab.