test(avatar): deleting a user avatar and file is atomic
The avatar must not be unset in the database if there is a failure to
remove the avatar file from storage (file or S3). The two operations
are wrapped in a transaction for that purpose and this test verifies
it is effective.

See 1be797faba301503e29db9de7eb32335a684464c Fix bug on avatar

(cherry picked from commit c139efb1e9e65998ad557159b0c93ddc871b6b59)