Rework the returned mail value
- Two behavior changes are noted with this patch, the display name is
now always quoted altough in some scenarios unnecessary it's a safety
precaution of Go. B encoding is used when certain characters are present
as they aren't 'legal' to be used as a display name and Q encoding would
still show them and B encoding needs to be used, this is now done by
Go's `address.String()`.
- Update and add new unit tests.