feat(cli): add `--keep-labels` flag to `forgejo actions register`
This commit adds a new flag, `--keep-labels`, to the runner
registration CLI command. If this flag is present and the runner being
registered already exists, it will prevent the runners' labels from
being reset.

In order to accomplish this, the signature of the `RegisterRunner`
function from the `models/actions` package has been modified so that the
labels argument can be nil. If it is, the part of the function that
updates the record will not change the runner.

Various tests have been added for this function, for the following
cases: new runner with labels, new runner without label, existing runner
with labels, existing runner without labels.

The flag has been added to the CLI command, the action function has been
updated to read the labels parameters through a separate function
(`getLabels`), and test cases for this function have been added.