feat(auth): add setting to disable password recovery
Having a password recovery system makes no sense for installations that
rely on external authentication for actual users and automation for
"special" users.

This commit adds a setting to the `[security]` section called
`DISABLE_PASSWORD_RECOVERY`. If this setting is set to `true`, the
following happen:

- the password recovery page returns a 404,
- the password recovery link no longer appears on the sign-in and
  account settings pages.

Tests for the setting itself and for the recovery page's GET handler
were added.