[SEC] Add `totp_recovery_code` as SSH command
- When a person loses access to their TOTP (e.g. phone wiped) and didn't
properly save their TOTP's scratch code they have to; they have to
rely on the instance admins to authenticate with them that it is
really their account, this can be a quite difficult and lengthy
process to safely verify this.
- Something that has come up in these situations is that such people
usually have a (verified) SSH key added to their account and could use
that to prove they are the owner of the account, by the possession of
such SSH key.
- This patch adds a new SSH command that streamlines this process to a
self-service, no involvement of instance admins are needed. This is
similar to how Gitlab does this:
https://gitlab.com/gitlab-org/gitlab-shell/-/merge_requests/74
- When the user wants to make use of this self-service, they simply have
to execute `ssh git@git.example.org totp_recovery_code`, they might need
to specify with `-i ~/.ssh/path_to_key` a verified key if this isn't the
default key, and confirm they want to regenerate the TOTP recovery code
and fill in their password. A mail notification will be sent to inform
the user that their TOTP recovery code has been regenerated via SSH,
this is done in the spirit of #4635.
- This reset function has a 5-minute rate limiter. To protect against
the scenario that this could be misused as a password brute force.
- Unit tests added.
- Integration tests added.
- Resolves #1933