Revert "Prevent allow/reject reviews on merged/closed PRs"
This reverts commit 4ed372af13a10e5a45464ac38f16c634707267c4.
This change from Gitea was not considered by the Forgejo UI team and there is a consensus that it feels like a regression.

The test which was added in that commit is kept and modified to test that reviews can successfully be submitted on closed and merged PRs.

Closes forgejo/design#11

(cherry picked from commit 65c2595f266060a2efc8ec4541c48ea61da2036e)