fix(dashboard): hide mirrors if pull mirrors off
Previously, the "Mirrors" tab would be visible in the dashboard
if mirrors were enabled to begin with. Although this feature
is still kind of a misfeature, given the possibility of pull
mirrors having existed and continuing to operate properly before
pull mirrors were disabled, this is deemed to be an edge case.

Even given the existence of this edge case, it is probably best
to make space for the "Collaborative" section anyway. This was
done instead of replacing "Mirrors" with "Collaborative" (which
currently gets hidden behind a menu), which would be perfect for
Codeberg, but this compromise also takes smaller instances that
are primarily or partially intended for mirroring repositories
into account.

The order of some elements in the DashboardRepoList.vue file
was also changed, for cosmetic reasons only.